requirements.txt
requests
django
djangorestframework
gunicorn
whitenoise
ipdb